+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 
 ## [Unreleased]
 
+## [0.1.17] - 2025-01-14
+
 ### Added
 
 - Enhanced `SaferAddUniqueConstraint` to support a `UniqueConstraint` with the